+++ plugins/netlify_cms.ts
@@ -53,8 +53,8 @@ export default function (userOptions?: Partial<Options>) {
     // Build the admin page
     site.addEventListener("afterRender", () => {
       const root = site.source.root!;
-      const config: Record<string, unknown> | undefined =
-        root.data[options.configKey] as Record<string, unknown> | undefined;
+      const config: Record<string, unknown> | undefined = root
+        .data[options.configKey] as Record<string, unknown> | undefined;
 
       if (!config) {
         throw new Exception("Missing configuration for Netlify CMS", {